cleaned up documentation a bit

+++ jarvis/pixels.h
@@ -6,24 +6,24 @@
 MkvsynthPixel getPixel               (uint8_t *payload, MkvsynthMetaData *metaData, int widthOffset, int heightOffset);
 void          putPixel               (MkvsynthPixel *pixel, uint8_t *payload, MkvsynthMetaData *metaData, int widthOffset, int heightOffset);
 
-int getRed                           (MkvsynthPixel *pixel);
-int getGreen                         (MkvsynthPixel *pixel);
-int getBlue                          (MkvsynthPixel *pixel);
+short getRed                         (MkvsynthPixel *pixel);
+short getGreen                       (MkvsynthPixel *pixel);
+short getBlue                        (MkvsynthPixel *pixel);
 MkvsynthPixel adjustRed              (MkvsynthPixel *pixel, double intensity);
 MkvsynthPixel adjustGreen            (MkvsynthPixel *pixel, double intensity);
 MkvsynthPixel adjustBlue             (MkvsynthPixel *pixel, double intensity);
 
-int getLuma                          (MkvsynthPixel *pixel);
-int getCb                            (MkvsynthPixel *pixel);
-int getCr                            (MkvsynthPixel *pixel);
+short getLuma                        (MkvsynthPixel *pixel);
+short getCb                          (MkvsynthPixel *pixel);
+short getCr                          (MkvsynthPixel *pixel);
 MkvsynthPixel adjustLuma             (MkvsynthPixel *pixel, double intensity);
 MkvsynthPixel adjustCb               (MkvsynthPixel *pixel, double intensity);
 MkvsynthPixel adjustCr               (MkvsynthPixel *pixel, double intensity);
 
-int getHue                           (MkvsynthPixel *pixel);
-int getLightness									   (MkvsynthPixel *pixel);
-int getSaturation                    (MkvsynthPixel *pixel);
-int getValue                         (MkvsynthPixel *pixel);
+short getHue                         (MkvsynthPixel *pixel);
+short getLightness                   (MkvsynthPixel *pixel);
+short getSaturation                  (MkvsynthPixel *pixel);
+short getValue                       (MkvsynthPixel *pixel);
 MkvsynthPixel adjustHue              (MkvsynthPixel *pixel, double degrees);
 MkvsynthPixel adjustLightness        (MkvsynthPixel *pixel, double intensity);
 MkvsynthPixel adjustSaturation       (MkvsynthPixel *pixel, double intensity);
